refactor: avoid testing implementation details
