+++ src/mixin.js
@@ -59,6 +59,7 @@ export default {
     if (!requested && this.$validator) {
       const Vue = this.$options._base; // the vue constructor.
       Vue.util.defineReactive(this.$validator, 'errors', this.$validator.errors);
+      Vue.util.defineReactive(this.$validator, 'fields', this.$validator.fields);
     }
 
     if (! this.$options.computed) {
@@ -68,11 +69,20 @@ export default {
     this.$options.computed[options.errorBagName || 'errors'] = function errorBagGetter () {
       return this.$validator.errors;
     };
-    this.$options.computed[options.fieldsBagName || 'fields'] = {
-      cache: false,
-      get: function fieldBagGetter () {
-        return this.$validator.flags;
-      }
+    this.$options.computed[options.fieldsBagName || 'fields'] = function fieldBagGetter () {
+      return this.$validator.fields.reduce((acc, field) => {
+        if (field.scope) {
+          if (!acc[`$${field.scope}`]) {
+            acc[`$${field.scope}`] = {};
+          }
+
+          acc[`$${field.scope}`][field.name] = field.flags;
+        }
+
+        acc[field.name] = field.flags;
+
+        return acc;
+      }, {});
     };
   },
   beforeDestroy () {
