+++ tests/directive/constrained-validation.js
@@ -1,0 +1,24 @@
+import { mount, createLocalVue } from '@vue/test-utils';
+import VeeValidate from '@/index';
+
+describe('HTML Constrained Validation API', () => {
+  test('sets HTML contstrained validation API messages', async () => {
+    const Vue = createLocalVue();
+
+    Vue.use(VeeValidate, {
+      validity: true
+    });
+    const wrapper = mount({
+      template: `<input type="text" name="email" v-validate="'email'">`
+    }, { localVue: Vue });
+
+    const input = wrapper.find('input');
+    input.element.value = 'not an email';
+    input.element.setCustomValidity = jest.fn();
+    await wrapper.vm.$validator.validate();
+    expect(input.element.setCustomValidity).toHaveBeenCalledTimes(1);
+    expect(input.element.setCustomValidity).toHaveBeenLastCalledWith(
+      wrapper.vm.$validator.errors.first('email')
+    );
+  });
+});
