--- tests/integration/components/Validity.vue
@@ -1,11 +1,0 @@
-<template>
-  <div>
-    <input type="text" name="email" v-validate="'email'">
-  </div>
-</template>
-
-<script>
-export default {
-  name: 'validity-test'
-};
-</script>
